backend/pyproject.toml
[project]
name = "normalizer-api"
version = "0.1.0"
description = "NORMALIZER backend API - cultural communication translation"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "openai>=1.50.0",
    "pyjwt[crypto]>=2.10.0",
    "cryptography>=44.0.0",
    "python-multipart>=0.0.18",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
